% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_VS_score.R
\name{cal_VS_score}
\alias{cal_VS_score}
\title{Function for calculating virtual stiffness score}
\usage{
cal_VS_score(eset = NULL, method = "ssgsea", ...)
}
\arguments{
\item{eset}{a matrix or dataframe (row: gene symbal id, col: sample id) (transcriptome data: CPM, TPM, RPKM, FPKM, etc.)}

\item{method}{method in GSVA, default 'ssgsea'}

\item{...}{other parameters in GSVA package}
}
\value{
a data frame
}
\description{
Function for calculating virtual stiffness score
}
\author{
Yi Xiong
}
